requirements.txt
python-dotenv
langchain_openai
langchain_core
langchain_community
streamlit
langchain
